requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
base58==2.1.1
charset-normalizer==3.3.2
ffmpeg-python==0.2.0
frozenlist==1.4.0
future==0.18.3
hashID==3.1.4
humanize==4.8.0
idna==3.4
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
multidict==6.0.4
nostril @ git+https://github.com/casics/nostril.git@fbc0c91249283a9fbc9036206391ce1138826fd3
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
plac==1.4.1
pluggy==1.3.0
prettytable==3.9.0
py-cord==2.4.1
pycord==0.1.1
pytesseract==0.3.10
pytest==7.4.3
python-dotenv==1.0.0
tabulate==0.9.0
wcwidth==0.2.10
yarl==1.9.2
